// services/chat-service/src/types/signaling.types.ts
export type DtlsParameters = any;
// eslint-disable-next-line @typescript-eslint/no-explicit-any
export type RtpCapabilities = any;
// eslint-disable-next-line @typescript-eslint/no-explicit-any
export type RtpParameters = any;
// ...
export const MEDIA_SERVER_API_BASE_URL = process.env.MEDIA_SERVER_URL || 'http://media_server_container:3001';

export const MediaServerApiEndpoints = {
  getRouterRtpCapabilities: (roomId: string) => `${MEDIA_SERVER_API_BASE_URL}/rooms/${roomId}/router-rtp-capabilities`,
  createTransport: (roomId: string) => `${MEDIA_SERVER_API_BASE_URL}/rooms/${roomId}/transports`,
  connectTransport: (roomId: string, transportId: string) => `${MEDIA_SERVER_API_BASE_URL}/rooms/${roomId}/transports/${transportId}/connect`,
  produce: (roomId: string, transportId: string) => `${MEDIA_SERVER_API_BASE_URL}/rooms/${roomId}/transports/${transportId}/produce`,
  consume: (roomId: string, transportId: string) => `${MEDIA_SERVER_API_BASE_URL}/rooms/${roomId}/transports/${transportId}/consume`,
  closeProducer: (roomId: string, producerId: string) => `${MEDIA_SERVER_API_BASE_URL}/rooms/${roomId}/producers/${producerId}/close`, // New, for cleanup
  closeTransport: (roomId: string, transportId: string) => `${MEDIA_SERVER_API_BASE_URL}/rooms/${roomId}/transports/${transportId}/close`, // New, for cleanup
};